kuksa/requirements.txt
paho-mqtt
kuksa-client
ditto-client @ git+https://github.com/eclipse-ditto/ditto-clients-python@b215e49b4ceaa7253ffc4b5c1db3968482a631bb